Sat, Jul 8, 2023, 3:46 PM +05:00

+++ half/halfs.c
@@ -20,9 +20,9 @@ float half(float bill, float tax, int tip)
 {
     bill = bill/2;
     tax = tax/2;
-    tip = tip/2;
+    tip = (float) tip/2;
     tax = (tax/100) * bill;
-    tip = (tip/100)*bill;
+    tip = (tip/100) * bill;
     printf("%f %f %i\n",bill,tax,tip);
     return 0.0;
 }
